requirements.txt
boto3==1.37.4
botocore==1.37.4
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.3
six==1.17.0
urllib3==2.3.0
